+++ themes/socialblue/gulpfile.js
@@ -4,8 +4,24 @@
 
 'use strict';
 
-var importOnce = require('node-sass-import-once'),
-    path = require('path');
+// ################################
+// Load Gulp and tools we will use.
+// ################################
+
+var importOnce  = require('node-sass-import-once'),
+    path        = require('path'),
+    gulp        = require('gulp'),
+    $           = require('gulp-load-plugins')(),
+    browserSync = require('browser-sync').create(),
+    del         = require('del'),
+    // gulp-load-plugins will report "undefined" error unless you load gulp-sass manually.
+    sass        = require('gulp-sass'),
+    kss         = require('kss'),
+    postcss     = require('gulp-postcss'),
+    autoprefixer= require('autoprefixer'),
+    mqpacker    = require('css-mqpacker'),
+    concat      = require('gulp-concat'),
+    runSequence = require('run-sequence');
 
 var options = {};
 
@@ -47,8 +63,8 @@ options.basetheme = {
 
 // Set the URL used to access the Drupal website under development. This will
 // allow Browser Sync to serve the website and update CSS changes on the fly.
-//options.drupalURL = '';
-options.drupalURL = 'http://social.dev:32769';
+options.drupalURL = '';
+// options.drupalURL = 'http://social.dev:32799';
 
 // Define the node-sass configuration. The includePaths is critical!
 options.sass = {
@@ -65,6 +81,7 @@ var sassFiles = [
   '!' + options.theme.components + '**/_*.scss'
 ];
 
+
 // On screen notification for errors while performing tasks
 var onError = function(err) {
   notify.onError({
@@ -73,12 +90,6 @@ var onError = function(err) {
     sound: "Beep"
   })(err);
   this.emit('end');
-};
-
-// Define icons source and destination paths
-options.icons = {
-  src   : options.theme.components + '01-base/icons/source/',
-  dest  : options.theme.build + 'icons/'
 };
 
 
@@ -99,98 +110,88 @@ options.styleGuide = {
   // The following paths are relative to the generated style guide.
   css: [
     // Base stylesheets
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'base.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'base.css'),
+    'kss-assets/base/base.css',
+    'kss-assets/css/base.css',
     // Atom stylesheets
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'alerts.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'alerts.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'badges.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'badges.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'button.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'button.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'cards.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'cards.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'form-controls.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'form-controls.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'labels.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'list-group.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'list-group.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'labels.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'close-icon.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'waves.css'),
+    'kss-assets/base/alerts.css',
+    'kss-assets/css/alerts.css',
+    'kss-assets/base/badges.css',
+    'kss-assets/css/badges.css',
+    'kss-assets/base/button.css',
+    'kss-assets/css/button.css',
+    'kss-assets/base/cards.css',
+    'kss-assets/css/cards.css',
+    'kss-assets/base/form-controls.css',
+    'kss-assets/css/form-controls.css',
+    'kss-assets/base/labels.css',
+    'kss-assets/css/list-group.css',
+    'kss-assets/base/list-group.css',
+    'kss-assets/css/labels.css',
+    'kss-assets/base/close-icon.css',
+    'kss-assets/css/waves.css',
     // Molecule stylesheets
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'dropdown.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'dropdown.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'file.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'file.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'form-elements.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'form-elements.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'datepicker.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'datepicker.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'input-groups.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'input-groups.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'password.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'password.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'timepicker.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'timepicker.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'media.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'mention.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'mention.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'breadcrumb.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'breadcrumb.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'nav-book.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'nav-book.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'nav-tabs.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'nav-tabs.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'navbar.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'navbar.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'pagination.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'pagination.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'popover.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'popover.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'teaser.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'teaser.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'tour.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'tour.css'),
+    'kss-assets/base/dropdown.css',
+    'kss-assets/css/dropdown.css',
+    'kss-assets/base/file.css',
+    'kss-assets/css/file.css',
+    'kss-assets/base/form-elements.css',
+    'kss-assets/css/form-elements.css',
+    'kss-assets/base/datepicker.css',
+    'kss-assets/css/datepicker.css',
+    'kss-assets/base/input-groups.css',
+    'kss-assets/css/input-groups.css',
+    'kss-assets/base/password.css',
+    'kss-assets/css/password.css',
+    'kss-assets/base/timepicker.css',
+    'kss-assets/css/timepicker.css',
+    'kss-assets/base/media.css',
+    'kss-assets/base/mention.css',
+    'kss-assets/css/mention.css',
+    'kss-assets/base/breadcrumb.css',
+    'kss-assets/css/breadcrumb.css',
+    'kss-assets/base/nav-book.css',
+    'kss-assets/css/nav-book.css',
+    'kss-assets/base/nav-tabs.css',
+    'kss-assets/css/nav-tabs.css',
+    'kss-assets/base/navbar.css',
+    'kss-assets/css/navbar.css',
+    'kss-assets/base/pagination.css',
+    'kss-assets/css/pagination.css',
+    'kss-assets/base/popover.css',
+    'kss-assets/css/popover.css',
+    'kss-assets/base/teaser.css',
+    'kss-assets/css/teaser.css',
+    'kss-assets/base/tour.css',
+    'kss-assets/css/tour.css',
     // Organisms stylesheets
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'comment.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'comment.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'hero.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'hero.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'meta.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'meta.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'offcanvas.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'offcanvas.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'site-footer.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'stream.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'stream.css'),
+    'kss-assets/base/comment.css',
+    'kss-assets/css/comment.css',
+    'kss-assets/base/hero.css',
+    'kss-assets/css/hero.css',
+    'kss-assets/base/meta.css',
+    'kss-assets/css/meta.css',
+    'kss-assets/base/offcanvas.css',
+    'kss-assets/css/offcanvas.css',
+    'kss-assets/css/site-footer.css',
+    'kss-assets/base/stream.css',
+    'kss-assets/css/stream.css',
     // Template stylesheets
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'layout.css'),
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'page-node.css'),
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'page-node.css'),
+    'kss-assets/base/layout.css',
+    'kss-assets/base/page-node.css',
+    'kss-assets/css/page-node.css',
     // Javascript stylesheets
-    path.relative(options.rootPath.styleGuide, options.basetheme.css + 'morrisjs.css'),
+    'kss-assets/base/morrisjs.css',
     // Styleguide stylesheets
-    path.relative(options.rootPath.styleGuide, options.theme.css + 'styleguide.css'),
-
+    'kss-assets/css/styleguide.css'
   ],
   js: [
-    path.relative(options.rootPath.styleGuide, options.theme.js + 'waves.min.js')
+    'kss-assets/js/waves.min.js'
   ],
   homepage: 'homepage.md',
   title: 'Style Guide for Open Social Blue'
 };
 
 
-// Define the paths to the JS files to lint.
-options.eslint = {
-  files  : [
-    options.rootPath.project + 'gulpfile.js',
-    options.theme.components + '**/*.js',
-    '!' + options.theme.components + '**/*.min.js',
-    '!' + options.theme.build + '**/*.js'
-  ]
-};
 
 // If your files are on a network share, you may want to turn on polling for
 // Gulp watch. Since polling is less efficient, we disable polling by default.
@@ -198,73 +199,70 @@ options.gulpWatchOptions = {};
 // options.gulpWatchOptions = {interval: 1000, mode: 'poll'};
 
 
-// ################################
-// Load Gulp and tools we will use.
-// ################################
-var gulp      = require('gulp'),
-  $           = require('gulp-load-plugins')(),
-  browserSync = require('browser-sync').create(),
-  del         = require('del'),
-  // gulp-load-plugins will report "undefined" error unless you load gulp-sass manually.
-  sass        = require('gulp-sass'),
-  kss         = require('kss'),
-  postcss     = require('gulp-postcss'),
-  autoprefixer= require('autoprefixer'),
-  mqpacker    = require('css-mqpacker'),
-  concat      = require('gulp-concat');
-
-// Must be defined after plugins are called.
 var sassProcessors = [
   autoprefixer({browsers: ['> 1%', 'last 2 versions']}),
   mqpacker({sort: true})
 ];
 
 
-// The default task.
-gulp.task('default', ['build']);
 
 // #################
-// Build everything.
+//
+// Compile the Sass
+//
 // #################
-gulp.task('build', ['styles:production', 'styleguide', 'lint']);
+//
+// This task will look for all scss files and run postcss and rucksack.
+// For performance review we will display the file sizes
+// Then the files will be stored in the assets folder
+// At the end we check if we should inject new styles in the browser
+// ===================================================
 
-
-gulp.task('styles', ['clean:css'], function () {
+gulp.task('styles', function () {
   return gulp.src(sassFiles)
-    .pipe($.sourcemaps.init())
     .pipe(sass(options.sass).on('error', sass.logError))
     .pipe($.plumber({ errorHandler: onError }) )
+    // run autoprefixer and media-query packer
     .pipe($.postcss(sassProcessors) )
+    // run rucksack @see https://simplaio.github.io/rucksack/
     .pipe($.rucksack() )
     .pipe($.rename({dirname: ''}))
     .pipe($.size({showFiles: true}))
-    .pipe($.sourcemaps.write('./'))
     .pipe(gulp.dest(options.theme.css))
     .pipe($.if(browserSync.active, browserSync.stream({match: '**/*.css'})));
 });
-
-gulp.task('styles:production', ['clean:css'], function () {
-  return gulp.src(sassFiles)
-    .pipe(sass(options.sass).on('error', sass.logError))
-    .pipe($.plumber({ errorHandler: onError }) )
-    .pipe($.postcss(sassProcessors) )
-    .pipe($.rucksack() )
-    .pipe($.rename({dirname: ''}))
-    .pipe($.size({showFiles: true}))
-    .pipe(gulp.dest(options.theme.css));
-});
-
 
 
 // ##################
 // Build style guide.
 // ##################
-gulp.task('styleguide', ['clean:styleguide', 'scripts-drupal'], function () {
+
+// Compile and copy the socialbase styles to the style guide
+gulp.task('styleguide-assets-base', function () {
+  return gulp.src(options.basetheme.css +'*.css')
+    .pipe(gulp.dest(options.rootPath.styleGuide + 'kss-assets/base/'))
+});
+
+// Compile and copy the subtheme assets to the style guide
+gulp.task('styleguide-assets', function () {
+  return gulp.src(options.theme.build +'**/*')
+    .pipe(gulp.dest(options.rootPath.styleGuide + 'kss-assets/'))
+});
+
+// Main styleguide task
+gulp.task('styleguide', ['clean:styleguide'], function () {
   return kss(options.styleGuide);
 });
 
+// Before deploying create a fresh build
+gulp.task('build-styleguide', function(done) {
+  runSequence('styleguide',
+    ['styleguide-assets-base', 'styleguide-assets'],
+    done);
+});
+
 // Debug the generation of the style guide with the --verbose flag.
-gulp.task('styleguide:debug', ['clean:styleguide'], function () {
+gulp.task('styleguide:debug', ['clean:styleguide', 'scripts-drupal'], function () {
   options.styleGuide.verbose = true;
   return kss(options.styleGuide);
 });
@@ -287,9 +285,12 @@ gulp.task('scripts-drupal', function() {
     .pipe( gulp.dest(options.rootPath.styleGuide + 'kss-assets/') );
 });
 
+
+
 // #########################
 // Lint Sass and JavaScript.
 // #########################
+
 gulp.task('lint', ['lint:sass', 'lint:js']);
 
 // Lint JavaScript.
@@ -322,10 +323,15 @@ gulp.task('lint:sass-with-fail', function () {
     .pipe($.sassLint.failOnError());
 });
 
+
+
 // ##############################
+//
 // Watch for changes and rebuild.
+//
 // ##############################
-gulp.task('watch', ['browser-sync', 'watch:lint-and-styleguide', 'watch:js']);
+
+gulp.task('watch', ['browser-sync', 'watch:styleguide', 'watch:js']);
 
 gulp.task('browser-sync', ['watch:css'], function () {
   if (!options.drupalURL) {
@@ -341,15 +347,26 @@ gulp.task('watch:css', ['styles'], function () {
   return gulp.watch(options.theme.components + '**/*.scss', ['styles']);
 });
 
-gulp.task('watch:lint-and-styleguide', ['styleguide'], function () {
+gulp.task('watch:styleguide', ['styleguide:debug', 'build-styleguide'], function () {
   return gulp.watch([
     options.basetheme.components + '**/*.scss',
     options.theme.components + '**/*.scss',
     options.basetheme.components + '**/*.twig',
     options.theme.components + '**/*.twig',
     options.theme.components + '**/*.md'
-  ], options.gulpWatchOptions, ['styleguide']);
+  ], options.gulpWatchOptions, 'build-styleguide');
 });
+
+// #################
+//
+// Minify JS
+//
+// #################
+//
+// First clean the JS folder, then search all components for js files.
+// Then compress the files, give them an explicit .min filename and
+// save them to the assets folder.
+// ===================================================
 
 gulp.task('watch:js', function () {
   return gulp.src(options.theme.components + '**/*.js')
@@ -361,9 +378,16 @@ gulp.task('watch:js', function () {
     .pipe(gulp.dest(options.theme.js));
 });
 
+
+
+
+
 // ######################
+//
 // Clean all directories.
+//
 // ######################
+
 gulp.task('clean', ['clean:css', 'clean:styleguide']);
 
 // Clean style guide files.
@@ -385,4 +409,59 @@ gulp.task('clean:css', function () {
 });
 
 
-// This gulpfile is based on the Drupal 8 Zen theme starterkit
+// ######################
+// Deploy
+// ######################
+//
+// We use rsync to publish the style guide online
+// ===================================================
+
+var rsync         = require('gulp-rsync'),
+    prompt        = require('gulp-prompt');
+
+try {
+  var deploy    = require('./deploy_config.json');
+} catch(error) {
+  console.log('Deploy config file missing');
+}
+
+// Before deploying create a fresh build
+gulp.task('deploy', ['build-styleguide'], function() {
+
+  // Default options for rsync
+  var rsyncConf = {
+    progress: true,
+    incremental: true,
+    emptyDirectories: true
+    //clean: true,
+  };
+
+  // Load the settings from our deploy_config.json file
+  // This file can be made from the deploy_config-example.json
+  // and should be in gitignore
+  rsyncConf.hostname = deploy.hostname; // hostname
+  rsyncConf.username = deploy.username; // ssh username
+  rsyncConf.destination = deploy.destination; // path where uploaded files go
+  rsyncConf.root = options.rootPath.styleGuide;
+
+  // Use gulp-rsync to sync the files
+  // and perform a final check before pushing
+  return gulp.src(options.rootPath.styleGuide + '/**/*')
+    .pipe(
+      prompt.confirm({
+        message: 'Heads Up! Are you SURE you want to push to PRODUCTION?',
+        default: false
+      })
+    )
+    .pipe(rsync(rsyncConf));
+
+});
+
+
+
+// ######################
+//
+// Default task (no watching)
+//
+// ######################
+gulp.task('default', ['styles', 'styleguide']);